refactor: extract data fetching logic into repository methods for RSC/tRPC sharing
- Add getScheduleListWithDefault method to ScheduleRepository
- Create getUserEventGroups service in @calcom/features/eventtypes/lib
- Move EventGroupBuilder, EventTypeGroupFilter, and related utils to @calcom/features
- Update RSC pages to use repository/service methods with unstable_cache
- Update tRPC handlers to use the same repository/service methods
- Re-export from @calcom/trpc for backward compatibility
