+++ packages/features/eventtypes/lib/getUserEventGroups/ProfilePermissionProcessor.ts
@@ -1,0 +1,11 @@
+import type { TeamPermissions } from "./permissionUtils";
+import { createProfilesWithPermissions, type EventTypeGroup, type ProfileWithPermissions } from "./transformUtils";
+
+export class ProfilePermissionProcessor {
+  processProfiles(
+    eventTypeGroups: EventTypeGroup[],
+    teamPermissionsMap: Map<number, TeamPermissions>
+  ): ProfileWithPermissions[] {
+    return createProfilesWithPermissions(eventTypeGroups, teamPermissionsMap);
+  }
+}
